chore: Update package.json scripts to include "pull" command
This commit updates the "package.json" file to include a new script command called "pull". The command uses "git reset --hard HEAD" and "git pull origin main" to reset the local repository to the latest commit and pull the latest changes from the remote repository. This addition improves the development workflow by providing an easy way to update the local codebase with the latest changes from the remote repository.